requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Vector database
qdrant-client==1.7.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
langchain==0.0.350
langchain-community==0.0.1

# Embeddings and LLM
sentence-transformers==2.2.2
transformers==4.35.2
torch>=1.9.0
ollama==0.1.7

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
requests==2.31.0
